requirements.txt
langchain
langchain-community
numpy
pandas
langchain-groq
python-dotenv
langchain-core
flask
-e .